하위뷰 분리, 주석정리
