+++ mohaemukzip/Sources/Feature/Recipe/Search/Views/VideoListView.swift
@@ -54,8 +54,3 @@ struct VideoListView: View {
     }
 }
 
-#Preview {
-    VideoListView(recipeVideoVM: RecipeVideoViewModel())
-        .environment(NavigationRouter())
-        .environment(SearchViewModel())
-}
